add campgrounds router
